backend/requirements.txt
Flask
Flask-SocketIO
langchain
pytz
deeplake
substrate-interface
python-dotenv
openai
firebase-admin